refactor: Boardコンテキストの値オブジェクト参照を整理
