+++ src/features/municipality/domain/value-objects/BoardAttributes.ts
@@ -1,22 +1,8 @@
-/**
- * 掲示板に関する列挙型定義
- *
- * ドメイン層で Prisma 依存を避けるための文字列リテラル型
- */
-
-export const BOARD_STATUS_VALUES = ["PENDING", "VERIFIED", "REJECTED"] as const;
-export type BoardStatus = (typeof BOARD_STATUS_VALUES)[number];
-
-export const TRUST_LEVEL_VALUES = [
-  "LEVEL_1",
-  "LEVEL_2",
-  "LEVEL_3",
-  "LEVEL_4",
-] as const;
-export type TrustLevel = (typeof TRUST_LEVEL_VALUES)[number];
-
-export const isBoardStatus = (value: string): value is BoardStatus =>
-  (BOARD_STATUS_VALUES as readonly string[]).includes(value);
-
-export const isTrustLevel = (value: string): value is TrustLevel =>
-  (TRUST_LEVEL_VALUES as readonly string[]).includes(value);
+export {
+  BOARD_STATUS_VALUES,
+  TRUST_LEVEL_VALUES,
+  isBoardStatus,
+  isTrustLevel,
+  type BoardStatus,
+  type TrustLevel,
+} from "@/shared/domain/board/BoardAttributes";
